@@ -1,12 +1,3 @@
-//buildscript {
-//    repositories {
-//        jcenter()
-//        google()
-//    }
-//    dependencies {
-//        classpath 'com.android.tools.build:gradle:3.1.4'
-//    }
-//}
 apply plugin: 'com.android.library'
 
 dependencies {
@@ -16,8 +7,7 @@ dependencies {
 android {
     namespace "org.jaudiotagger"
 
-    compileSdkVersion 27
-    //buildToolsVersion"27.0.3"
+    compileSdk 35
 
     sourceSets {
         main {
