@@ -1,12 +1,3 @@
-//buildscript {
-//    repositories {
-//        jcenter()
-//        google()
-//    }
-//    dependencies {
-//        classpath 'com.android.tools.build:gradle:3.1.4'
-//    }
-//}
 apply plugin: 'com.android.library'
 
 dependencies {
@@ -19,8 +10,8 @@ dependencies {
 android {
     namespace "com.sothree.slidinguppanel.library"
 
-    compileSdkVersion 27
-    //buildToolsVersion"27.0.3"
+    compileSdk 35
+
     sourceSets {
         main {
             manifest.srcFile 'AndroidManifest.xml'
