@@ -20,10 +20,9 @@ dependencies {
 android {
     namespace "kr.kymedia.karaoke.api"
 
-    publishNonDefault true
+    compileSdk 35
 
-    compileSdkVersion 27
-    //buildToolsVersion"27.0.3"
+    publishNonDefault true
 
     //useLibrary 'org.apache.http.legacy'
 
