@@ -1,12 +1,3 @@
-//buildscript {
-//    repositories {
-//        jcenter()
-//        google()
-//    }
-//    dependencies {
-//        classpath 'com.android.tools.build:gradle:3.1.4'
-//    }
-//}
 apply plugin: 'com.android.library'
 
 dependencies {
@@ -21,10 +12,9 @@ dependencies {
 android {
     namespace "kr.kymedia.karaoke.play"
 
-    publishNonDefault true
+    compileSdk 35
 
-    compileSdkVersion 27
-    //buildToolsVersion"27.0.3"
+    publishNonDefault true
 
     sourceSets {
         main {
